requirements_agentic.txt
# FastAPI for backend
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.7.0

# Claude API integration
anthropic>=0.42.0

# Environment variables
python-dotenv>=1.0.0

# CORS middleware
starlette>=0.36.0 